backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.40

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==5.0.0

# ── Validation & Config ───────────────────────────────────────────────────────
pydantic==2.11.2
pydantic-settings==2.8.1

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── AI / ML ───────────────────────────────────────────────────────────────────
scikit-learn==1.6.1
pandas==2.2.3
joblib==1.4.2
requests==2.32.3

# ── Testing (dev only — exclude from prod image) ──────────────────────────────
pytest==8.3.5
httpx==0.28.1
